Improved and refactored both user and admin interfaces
There was a problem when displaying of users after the winners
were generated. It didn't give you the current list of winners,
but it would just show the SelectRandomWinners button
that would use the current set of winners and keep incrementing
winners from the existing.

Solution: to make winners a more global property that is available
thorughout the app, and display the list of winners if they
have been already selected.